requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiohttp-swagger3==0.7.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
fastjsonschema==2.16.3
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pypng==0.20220715.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
rfc3339-validator==0.1.4
six==1.16.0
typing_extensions==4.6.3
yarl==1.9.2
